// src/client/types.gen.ts
export type RoleRead = {
    /**
     * Name
     *
     * Name of the role within the system.
     */
    name: 'admin' | 'manager' | 'user';
    /**
     * Id
     */
    id: number;
};
// ...
export type UserRead = {
    /**
     * Username
     */
    username: string;
    /**
     * Name
     */
    name?: string | null;
    /**
     * Profile Photo Url
     */
    profile_photo_url?: string | null;
    /**
     * Is Active
     */
    is_active?: boolean;
    /**
     * Id
     */
    id: string;
    /**
     * Email
     */
    email: string;
    role: RoleRead;
    /**
     * Google Id
     */
    google_id?: string | null;
    /**
     * Google Auth Active
     */
    google_auth_active: boolean;
    /**
     * Last Login
     */
    last_login?: Date | null;
    /**
     * Created At
     */
    created_at: Date;
    /**
     * Updated At
     */
    updated_at: Date;
    /**
     * Is Deleted
     */
    is_deleted: boolean;
};